feat(core): move the floating menu plugin into core and drop the extension-block-menu peer dependency from all framework wrappers
