+++ packages/angular/package.json
@@ -16,8 +16,7 @@
   "peerDependencies": {
     "@angular/core": ">=17.1.0",
     "@angular/forms": ">=17.1.0",
-    "@domternal/core": ">=0.8.0",
-    "@domternal/extension-block-menu": ">=0.8.0"
+    "@domternal/core": ">=0.8.0"
   },
   "files": [
     "dist"
@@ -33,7 +32,6 @@
     "@angular/forms": "^21.2.4",
     "@angular/platform-browser": "^21.2.4",
     "@domternal/core": "workspace:*",
-    "@domternal/extension-block-menu": "workspace:*",
     "ng-packagr": "^21.0.0",
     "typescript": "~5.9.3"
   },
