+++ packages/core/src/Editor.ts
@@ -162,6 +162,8 @@ export class Editor extends EventEmitter<EditorEvents> {
   get isEditable(): boolean {
     // The view is unset while EditorView's constructor runs the initial draw
     // (plugin decoration props may call this), so fall back to the option.
+    // The declared type says `view` is always set; mid-construction it is not.
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     return this.view ? this.view.editable : (this.options.editable ?? true);
   }
