+++ packages/react/src/DomternalFloatingMenu.tsx
@@ -13,19 +13,17 @@ import {
 import {
   PluginKey,
   FloatingMenuController,
+  createFloatingMenuPlugin,
   defaultIcons,
 } from '@domternal/core';
 import type {
   Editor,
   FloatingMenuItem,
   FloatingMenuItemsOverride,
-  IconSet,
-} from '@domternal/core';
-import { createFloatingMenuPlugin } from '@domternal/extension-block-menu';
-import type {
   FloatingMenuKeymap,
   FloatingMenuOptions,
-} from '@domternal/extension-block-menu';
+  IconSet,
+} from '@domternal/core';
 import { useCurrentEditor } from './EditorContext.js';
 
 export interface DomternalFloatingMenuProps {
