+++ packages/vanilla/package.json
@@ -19,12 +19,10 @@
     "dist"
   ],
   "peerDependencies": {
-    "@domternal/core": ">=0.8.0",
-    "@domternal/extension-block-menu": ">=0.8.0"
+    "@domternal/core": ">=0.8.0"
   },
   "devDependencies": {
     "@domternal/core": "workspace:*",
-    "@domternal/extension-block-menu": "workspace:*",
     "tsup": "^8.5.1",
     "typescript": "~5.9.3"
   },
