+++ packages/vue/package.json
@@ -20,12 +20,10 @@
   ],
   "peerDependencies": {
     "vue": ">=3.3",
-    "@domternal/core": ">=0.8.0",
-    "@domternal/extension-block-menu": ">=0.8.0"
+    "@domternal/core": ">=0.8.0"
   },
   "devDependencies": {
     "@domternal/core": "workspace:*",
-    "@domternal/extension-block-menu": "workspace:*",
     "vue": "^3.5.13",
     "tsup": "^8.5.1",
     "typescript": "~5.9.3"
